Fix bottom-board rotations
I think subtracting the part rotation when it is on the bottom of the board is wrong. It has no effect on parts with 0 or 180 degree corrections, and they are placed incorrectly on JLC's part viewer. I've changed it to simply add 180 degrees. This shouldn't affect parts with 90/270 degree corrections, but will make parts with 0 or 180 degree corrections place correctly.